Do not put migrations under their own root
Since the migrations are not displayed on the dirtributions, there is not need to organize them under their own root to reduce visual clutter.  Having each migration follow the same path as all other distributions makes each easier to find in the absence of a link displayed on the distributions web page.  It also avoids complicating the distribution deployment scripts and allows each migration distribution to be treated the same as any other distribution.